Pvekxx

+++ backend/src/signup/signup.service.ts
@@ -87,6 +87,7 @@ export class SignupService {
   async createJwtToken(userInfo: CreateSignupDto) {
     const user = userInfo.data;
 
+
     const jwt = this.jwtService.sign(user);
 
     return jwt;
@@ -112,7 +113,7 @@ export class SignupService {
       }
 
       const mailOptions = {
-        from: 'dkswndgus0506@naver.com',
+        from: process.env.DEVELOP_EMAIL,
         to: createAuthNum.data.email,
         subject: '약 알람 bbiyakbbiyak 앱 회원 인증 관련 이메일입니다.',
         html: '<h1>인증번호를 입력해주세요 \n\n\n\n\n\n</h1>' + number,
@@ -193,31 +194,12 @@ export class SignupService {
     }
   }
 
-  async deleteUser(checkToken: string) {
-    const token = checkToken.split(' ').pop();
+  async deleteUser(userId: number) {
 
-    const tokenPayload = await this.jwtService.verifyAsync(token);
 
     try {
-      if (
-        tokenPayload.isOAuthUser === true &&
-        tokenPayload.email_verified === true
-      ) {
-        await this.userSignupLogic.destroy({
-          where: {
-            email: tokenPayload.email,
-            isOAuthUser: tokenPayload.isOAuthUser,
-          },
-        });
-        console.log('Oauth 유저 회원 탈퇴함');
-        return;
-      } else if (tokenPayload.email && tokenPayload.isOAuthUser === false) {
-        await this.userSignupLogic.destroy({
-          where: { email: tokenPayload.email, phone: tokenPayload.phone },
-        });
-        console.log('개인회원 유저 회원 탈퇴함');
-        return;
-      }
+      await this.userSignupLogic.destroy({ where: { id: userId } })
+
     } catch (error) {
       throw new HttpException('회원탈퇴 실패', HttpStatus.FORBIDDEN);
     }
